+++ ia1/NodeFactUnary.java
@@ -1,12 +1,24 @@
+//class to handle the unary operation
 public class NodeFactUnary extends NodeFact {
 
     private NodeFact fact;
 
+    /**
+     * constructor to initialize the variables
+     * @param fact
+     */
     public NodeFactUnary(NodeFact fact) {
         this.fact= fact;
     }
 
-    public int eval(Environment env) throws EvalException {
+    /**
+     * takes in a environemnt object
+     * returns - 1 times the return of the fact
+     * @param env
+     * @return double value
+     * @throws EvalException
+     */
+    public double eval(Environment env) throws EvalException {
         return -1*fact.eval(env);
     }
